Add more structure rook code
This PR aims to structure rook code as it was done for kubernetes
The RookBase class is inherited by various implementation of
rook (upstream, ses etc...) and imported as RookCluster
